Extract map error middleware
