+++ lib/cartodb/controllers/map/named.js
@@ -1,4 +1,3 @@
-const _ = require('underscore');
 const ResourceLocator = require('../../models/resource-locator');
 const cleanUpQueryParams = require('../../middleware/clean-up-query-params');
 const credentials = require('../../middleware/credentials');
@@ -16,6 +15,7 @@ const lastUpdatedTimeLayergroup = require('./middlewares/last-updated-time-layer
 const layerStats = require('./middlewares/layer-stats');
 const layergroupIdHeader = require('./middlewares/layergroup-id-header');
 const layergroupMetadata = require('./middlewares/layergroup-metadata');
+const mapError = require('./middlewares/map-error');
 const sendResponse = require('../../middleware/send-response');
 const NamedMapMapConfigProvider = require('../../models/mapconfig/provider/named-map-provider');
 const CreateLayergroupMapConfigProvider = require('../../models/mapconfig/provider/create-layergroup-provider');
@@ -119,7 +119,7 @@ NamedMapController.prototype.composeInstantiateTemplateMiddleware = function () 
         layergroupIdHeader(this.templateMaps ,useTemplateHash),
         layergroupMetadata(this.layergroupMetadata, includeQuery),
         sendResponse(),
-        augmentError({ label, addContext })
+        mapError({ label, addContext })
     ];
 };
 
@@ -221,40 +221,4 @@ function instantiateLayergroup (mapBackend, userLimitsApi, pgConnection, affecte
             next();
         });
     };
-}
-
-function augmentError (options) {
-    const { addContext = false, label = 'MAPS CONTROLLER' } = options;
-
-    return function augmentErrorMiddleware (err, req, res, next) {
-        req.profiler.done('error');
-        const { mapConfig } = res.locals;
-
-        if (addContext) {
-            err = Number.isFinite(err.layerIndex) ? populateError(err, mapConfig) : err;
-        }
-
-        err.label = label;
-
-        next(err);
-    };
-}
-
-function populateError(err, mapConfig) {
-    var error = new Error(err.message);
-    error.http_status = err.http_status;
-
-    if (!err.http_status && err.message.indexOf('column "the_geom_webmercator" does not exist') >= 0) {
-        error.http_status = 400;
-    }
-
-    error.type = 'layer';
-    error.subtype = err.message.indexOf('Postgis Plugin') >= 0 ? 'query' : undefined;
-    error.layer = {
-        id: mapConfig.getLayerId(err.layerIndex),
-        index: err.layerIndex,
-        type: mapConfig.layerType(err.layerIndex)
-    };
-
-    return error;
 }
